move extrapolated components to different dir
